pdf_de_project/requirements.txt
pdfplumber
pymupdf
snowflake-connector-python
dbt-core>=1.7
dbt-snowflake>=1.7
pandas
python-dotenv
requests